--- src/main/java/br/com/startUp/Repository/PessoaRepository.java
@@ -1,10 +1,0 @@
-package br.com.startUp.Repository;
-
-import br.com.startUp.model.Pessoa;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PessoaRepository extends JpaRepository<Pessoa, Long> {
-
-}
