--- src/main/java/br/com/startUp/Repository/ProfileRepository.java
@@ -1,4 +1,0 @@
-package br.com.startUp.Repository;
-
-public interface ProfileRepository {
-}
