+++ src/main/java/br/com/startUp/StartUpApplication.java
@@ -7,6 +7,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class StartUpApplication {
 
 	public static void main(String[] args) {
+
 		SpringApplication.run(StartUpApplication.class, args);
 	}
