+++ src/main/java/br/com/startUp/controller/EnderecoController.java
@@ -1,7 +1,8 @@
-package com.br.curso.startup.controller;
+package br.com.startUp.controller;
 
 import java.util.List;
 
+import br.com.startUp.service.EnderecoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,8 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.br.curso.startup.dto.EnderecoDTO;
-import com.br.curso.startup.service.EnderecoService;
+import br.com.startUp.dto.EnderecoDTO;
 
 @RestController
 @RequestMapping("/pessoa")
