+++ src/main/java/br/com/startUp/model/Endereco.java
@@ -1,85 +1,4 @@
 package br.com.startUp.model;
 
-import javax.persistence.*;
-
-@Entity
-@Table(name = "endereco")
 public class Endereco {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id_endereco")
-    private Integer id;
-    @Column(name = "logradouro")
-    private String logradouro;
-    @Column(name = "cidade")
-    private String cidade;
-    @Column(name = "estado")
-    private String estado;
-    @Column(name = "complemento")
-    private String complemento;
-
-    @ManyToOne
-    @JoinColumn(name = "id_pessoa", nullable = true)
-    private Pessoa pessoa;
-
-    public Endereco(){
-
-    }
-
-    public Endereco(String logradouro, String cidade, String estado, String complemento, Pessoa pessoa) {
-        this.logradouro = logradouro;
-        this.cidade = cidade;
-        this.estado = estado;
-        this.complemento = complemento;
-        this.pessoa = pessoa;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getLogradouro() {
-        return logradouro;
-    }
-
-    public void setLogradouro(String logradouro) {
-        this.logradouro = logradouro;
-    }
-
-    public String getCidade() {
-        return cidade;
-    }
-
-    public void setCidade(String cidade) {
-        this.cidade = cidade;
-    }
-
-    public String getEstado() {
-        return estado;
-    }
-
-    public void setEstado(String estado) {
-        this.estado = estado;
-    }
-
-    public String getComplemento() {
-        return complemento;
-    }
-
-    public void setComplemento(String complemento) {
-        this.complemento = complemento;
-    }
-
-    public Pessoa getPessoa() {
-        return pessoa;
-    }
-
-    public void setPessoa(Pessoa pessoa) {
-        this.pessoa = pessoa;
-    }
 }
