+++ src/main/java/br/com/startUp/model/Profilie.java
@@ -1,4 +1,4 @@
 package br.com.startUp.model;
 
-public class Pessoa {
+public class Profilie {
 }
