+++ src/main/java/br/com/startUp/repository/EnderecoRepository.java
@@ -1,4 +1,4 @@
-package com.br.curso.startup.repository;
+package br.com.startUp.repository;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.br.curso.startup.dto.EnderecoDTO;
+import br.com.startUp.dto.EnderecoDTO;
 
 @Repository
 public interface EnderecoRepository extends JpaRepository<EnderecoDTO, Integer> {
