+++ src/main/java/br/com/startUp/service/EnderecoService.java
@@ -1,4 +1,4 @@
-package com.br.curso.startup.service;
+package br.com.startUp.service;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.br.curso.startup.dto.EnderecoDTO;
-import com.br.curso.startup.repository.EnderecoRepository;
+import br.com.startUp.dto.EnderecoDTO;
+import br.com.startUp.repository.EnderecoRepository;
 
 @Service
 public class EnderecoService {
