--- src/main/java/com/br/curso/startup/StratupApplication.java
@@ -1,14 +1,0 @@
-package com.br.curso.startup;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StratupApplication {
-
-	public static void main(String[] args) {
-
-		SpringApplication.run(StratupApplication.class, args);
-	}
-
-}
