--- src/main/java/com/br/curso/startup/dto/ProfileDTO.java
@@ -1,69 +1,0 @@
-package com.br.curso.startup.dto;
-
-import javax.annotation.Generated;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "tb_profile")
-public class ProfileDTO {
-    
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Integer id;
-    @Column(name = "youtube_channel")
-    private String youtubeChannel;
-    @Column(name = "facebook")
-    private String facebook;
-    @Column(name = "linkedin")
-    private String linkedin;
-
-    public ProfileDTO(){
-
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getYoutubeChannel() {
-        return youtubeChannel;
-    }
-
-    public void setYoutubeChannel(String youtubeChannel) {
-        this.youtubeChannel = youtubeChannel;
-    }
-
-    public String getFacebook() {
-        return facebook;
-    }
-
-    public void setFacebook(String facebook) {
-        this.facebook = facebook;
-    }
-
-    public String getLinkedin() {
-        return linkedin;
-    }
-
-    public void setLinkedin(String linkedin) {
-        this.linkedin = linkedin;
-    }
-
-    public ProfileDTO(String youtubeChannel, String facebook, String linkedin) {
-        this.youtubeChannel = youtubeChannel;
-        this.facebook = facebook;
-        this.linkedin = linkedin;
-    }
-
-    
-}
