--- src/main/java/com/br/curso/startup/model/Profilie.java
@@ -1,4 +1,0 @@
-package com.br.curso.startup.model;
-
-public class Profilie {
-}
