--- src/test/java/com/br/curso/startup/StratupApplicationTests.java
@@ -1,13 +1,0 @@
-package com.br.curso.startup;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class StratupApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
